backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
sqlite3
PyMuPDF==1.23.14
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0